Remove email address from the UserSummaryDTO, relevant to ucam-cl-dtg/isaac-app#794
# This increases user email privacy from 'hidden, but still disclosed' to 'not disclosed unless necessary'.
