+++ src/main/java/uk/ac/cam/cl/dtg/segue/api/managers/UserAssociationManager.java
@@ -278,7 +278,6 @@ public class UserAssociationManager {
             userRequested.setAuthorisedFullAccess(true);
         } else {
             userRequested.setAuthorisedFullAccess(false);
-            userRequested.setEmail(null);
         }
         return userRequested;
     }
